refactor: extract flow-narrowing into src/builder/narrowing.rs
Move guard recognition + span-scoped emission out of the builder.rs bulk into
a dedicated child module. It stays in the Build layer (rule #1: still driven by
build(), still the single tree-sitter consumer) — a CHILD module so its
impl Builder methods keep direct private-field access without pub(crate) leakage.
Generic helpers (point_lt, raw_mid_op, raw_leading_op, connector_keyword_between,
node_to_span) stay in builder.rs and are reached via super::/crate::cst.

Generalize the layering test's directory scan so src/builder/ contributes its
.rs files to the Build layer (a submodule can't dodge the import DAG). Pure
move, behavior-preserving: 956 tests green, no warnings.
